requirements.txt
flask
flask-socketio
ultralytics
opencv-python-headless
python-telegram-bot
requests
eventlet
# For development purposes, you might want to include: